backend/fastapi/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   starlette
    #   watchfiles
click==8.1.7
    # via uvicorn
dnspython==2.4.2
    # via
    #   email-validator
    #   pymongo
email-validator==2.1.0.post1
    # via pydantic
exceptiongroup==1.2.0
    # via anyio
fastapi==0.108.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   email-validator
motor==3.3.2
    # via -r requirements.in
pydantic[email]==2.5.3
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.14.6
    # via pydantic
pydantic-settings==2.1.0
    # via -r requirements.in
pymongo==4.6.1
    # via motor
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.1
    # via uvicorn
sniffio==1.3.0
    # via anyio
starlette==0.32.0.post1
    # via fastapi
typing-extensions==4.9.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
uvicorn[standard]==0.25.0
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
